Added `collectionKey` option for HasMany relations
If you set `collectionKey` to `true` on a HasMany relation (default), the relation's
`key` attribute will be used to create a reference to the RelationalModel instance
from the generated collection. Alternatively, if you set `collectionKey` to a
string, it will use that string as the reference to the RelationalModel, rather
than the relation's `key` attribute. If you set `collectionKey` to false (or any
falsy value) this reference will not be created.