Add support for submodels.
Backbone.RelationalModel.extend options:
partOfSupermodel: `true` if this model should use the same model collection and relations as its superclass.
submodelTypeAttribute: attribute that should be checked to determine the type of object to build when attributes are passed in. Default: `type`
submodelType: value of attribute defined using submodelTypeAttribute that will cause this type of object to be built.