SubModelTypes as model constructor or string
Of all the options that allow you to provide a Model definition, this is one of few that for some reason only allows you to provide it as string. 

This change puts this option in line with the rest of the similar options, no longer forcing you to use model scopes. 

Notations such as: 
subModelTypes: {
    "fade": FadeActionModel
}
were unsupported before.